PHP Server Update
- Added some changes to the code so it works on our AWS server
- Moving forward we must stop using relative file names for 'requires'
and 'includes'. They easily break. Instead dynamically determine the
base directory "/var/www/html/senior/tutor_web_app" and base ALL
requires off of that. Or it will break again. Examples can be seen in
this commit.
- The  OAuth server is not fully working still. Working with php is
proving difficult. We may want to consider switching to Java (lots of
the code can be recycled).
- '/public/api_index.php' is working but '/public/api_login.php' is
still having issues with OAuth - need help with this.